+++ xdoctest/directive.py
@@ -70,11 +70,10 @@ arguments allow you to condition on:
 CommandLine:
     python -m xdoctest.directive __doc__
 
+The following example shows how the ``+SKIP`` directives may be used to
+bypass certain places in the code.
 
 Example:
-    The following example shows how the ``+SKIP`` directives may be used to
-    bypass certain places in the code.
-
     >>> # An inline directive appears on the same line as a command and
     >>> # only applies to the current line.
     >>> raise AssertionError('this will not be run (a)')  # xdoctest: +SKIP
@@ -98,10 +97,10 @@ Example:
     >>> # xdoctest: -SKIP
     >>> print('This line will print: (F)')
 
-Example:
-    This next examples illustrates how to use the advanced ``+REQURIES()``
-    directive. Note, the REQUIRES and SKIP states are independent.
+This next examples illustrates how to use the advanced ``+REQURIES()``
+directive. Note, the REQUIRES and SKIP states are independent.
 
+Example:
     >>> import sys
     >>> count = 0
     >>> # xdoctest: +REQUIRES(WIN32)
@@ -338,10 +337,31 @@ class Directive(utils.NiceRepr):
         Yeilds:
             Directive: directive: the parsed directives
 
-        Notes:
+        Note:
             The original ``doctest`` module sometimes yeilded false positives for a
             directive pattern. Because ``xdoctest`` is parsing the text, this issue
             does not occur.
+
+        Example:
+            >>> from xdoctest.directive import Directive, RuntimeState
+            >>> state = RuntimeState()
+            >>> assert len(state['REQUIRES']) == 0
+            >>> extracted1 = list(Directive.extract('# xdoctest: +REQUIRES(CPYTHON)'))
+            >>> extracted2 = list(Directive.extract('# xdoctest: +REQUIRES(PYPY)'))
+            >>> print('extracted1 = {!r}'.format(extracted1))
+            >>> print('extracted2 = {!r}'.format(extracted2))
+            >>> effect1 = extracted1[0].effects()[0]
+            >>> effect2 = extracted2[0].effects()[0]
+            >>> print('effect1 = {!r}'.format(effect1))
+            >>> print('effect2 = {!r}'.format(effect2))
+            >>> assert effect1.value == 'CPYTHON'
+            >>> assert effect2.value == 'PYPY'
+            >>> # At least one of these will not be satisfied
+            >>> assert effect1.action == 'set.add' or effect2.action == 'set.add'
+            >>> state.update(extracted1)
+            >>> state.update(extracted2)
+            >>> print('state = {!r}'.format(state))
+            >>> assert len(state['REQUIRES']) > 0
 
         Example:
             >>> from xdoctest.directive import Directive
@@ -373,6 +393,7 @@ class Directive(utils.NiceRepr):
             <Directive(+REQUIRES(module:foo, module:bar))>
 
         Example:
+            >>> from xdoctest.directive import Directive, RuntimeState
             >>> any(Directive.extract(' # xdoctest: skip'))
             True
             >>> any(Directive.extract(' # badprefix: not-a-directive'))
@@ -409,7 +430,7 @@ class Directive(utils.NiceRepr):
             return '{}{}'.format(prefix, self.name)
 
     def _unpack_args(self, num):
-        warnings.warning('Deprecated and will be removed', DeprecationWarning)
+        warnings.warn('Deprecated and will be removed', DeprecationWarning)
         nargs = self.args
         if len(nargs) != 1:
             raise TypeError(
@@ -418,7 +439,7 @@ class Directive(utils.NiceRepr):
         return self.args
 
     def effect(self, argv=None, environ=None):
-        warnings.warning('Deprecated use effects', DeprecationWarning)
+        warnings.warn('Deprecated use effects', DeprecationWarning)
         effects = self.effects(argv=argv, environ=environ)
         if len(effects) > 1:
             raise Exception('Old method cannot hanldle multiple effects')
@@ -572,6 +593,7 @@ def _is_requires_satisfied(arg, argv=None, environ=None):
         bool: flag - True if the requirement is met
 
     Example:
+        >>> from xdoctest.directive import *  # NOQA
         >>> _is_requires_satisfied('PY2', argv=[])
         >>> _is_requires_satisfied('PY3', argv=[])
         >>> _is_requires_satisfied('cpython', argv=[])
@@ -651,12 +673,12 @@ def _is_requires_satisfied(arg, argv=None, environ=None):
         else:
             raise ValueError('Too many expr_parts={}'.format(expr_parts))
     elif arg_lower in SYS_PLATFORM_TAGS:
-        flag = sys.platform.startswith(arg_lower)
+        flag = sys.platform.lower().startswith(arg_lower)
     elif arg_lower in OS_NAME_TAGS:
         flag = os.name.startswith(arg_lower)
     elif arg_lower in PY_IMPL_TAGS:
         import platform
-        flag = platform.python_implementation().startswith(arg_lower)
+        flag = platform.python_implementation().lower().startswith(arg_lower)
     elif arg_lower in PY_VER_TAGS:
         if sys.version_info[0] == 2:  # nocover
             flag = arg_lower == 'py2'